Create common widgets for game app bar
