+++ lib/src/view/game/lobby_game_screen.dart
@@ -8,18 +8,13 @@ import 'package:lichess_mobile/src/model/game/game_repository_providers.dart';
 import 'package:lichess_mobile/src/model/lobby/lobby_game.dart';
 import 'package:lichess_mobile/src/model/lobby/game_seek.dart';
 import 'package:lichess_mobile/src/model/user/user_repository_providers.dart';
-import 'package:lichess_mobile/src/widgets/adaptive_bottom_sheet.dart';
-import 'package:lichess_mobile/src/widgets/buttons.dart';
 import 'package:lichess_mobile/src/widgets/platform.dart';
 import 'package:lichess_mobile/src/utils/immersive_mode.dart';
 import 'package:lichess_mobile/src/utils/wakelock.dart';
-import 'package:lichess_mobile/src/utils/l10n_context.dart';
 
 import 'game_body.dart';
-import 'ping_rating.dart';
+import 'game_common_widgets.dart';
 import 'game_loading_board.dart';
-import 'game_settings.dart';
-import 'game_screen_providers.dart';
 
 /// Screen for games created from the lobby.
 ///
@@ -74,28 +69,8 @@ class _GameScreenState extends ConsumerState<LobbyGameScreen>
   }
 
   Widget _androidBuilder(BuildContext context) {
-    final isPlayableAsync = ref.watch(lobbyGameIsPlayableProvider(widget.seek));
-    const pingRating = Padding(
-      padding: EdgeInsets.symmetric(horizontal: 16.0, vertical: 18.0),
-      child: PingRating(size: 24.0),
-    );
     return Scaffold(
-      appBar: AppBar(
-        leading: isPlayableAsync.maybeWhen<Widget?>(
-          data: (isPlayable) => isPlayable ? pingRating : null,
-          orElse: () => pingRating,
-        ),
-        title: _GameTitle(seek: widget.seek),
-        actions: [
-          SettingsButton(
-            onPressed: () => showAdaptiveBottomSheet<void>(
-              context: context,
-              isScrollControlled: true,
-              builder: (_) => GameSettings(seek: widget.seek),
-            ),
-          ),
-        ],
-      ),
+      appBar: GameAppBar(seek: widget.seek),
       body: _LoadLobbyGame(
         seek: widget.seek,
         whiteClockKey: _whiteClockKey,
@@ -105,28 +80,8 @@ class _GameScreenState extends ConsumerState<LobbyGameScreen>
   }
 
   Widget _iosBuilder(BuildContext context) {
-    final isPlayableAsync = ref.watch(lobbyGameIsPlayableProvider(widget.seek));
-    const pingRating = Padding(
-      padding: EdgeInsets.symmetric(horizontal: 16.0, vertical: 12.0),
-      child: PingRating(size: 24.0),
-    );
-
     return CupertinoPageScaffold(
-      navigationBar: CupertinoNavigationBar(
-        padding: const EdgeInsetsDirectional.only(end: 16.0),
-        leading: isPlayableAsync.maybeWhen<Widget?>(
-          data: (isPlayable) => isPlayable ? pingRating : null,
-          orElse: () => pingRating,
-        ),
-        middle: _GameTitle(seek: widget.seek),
-        trailing: SettingsButton(
-          onPressed: () => showAdaptiveBottomSheet<void>(
-            context: context,
-            isScrollControlled: true,
-            builder: (_) => GameSettings(seek: widget.seek),
-          ),
-        ),
-      ),
+      navigationBar: GameCupertinoNavBar(seek: widget.seek),
       child: _LoadLobbyGame(
         seek: widget.seek,
         whiteClockKey: _whiteClockKey,
@@ -174,31 +129,6 @@ class _LoadLobbyGame extends ConsumerWidget {
           child: LoadGameError(),
         );
       },
-    );
-  }
-}
-
-class _GameTitle extends ConsumerWidget {
-  const _GameTitle({
-    required this.seek,
-  });
-
-  final GameSeek seek;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final mode =
-        seek.rated ? ' • ${context.l10n.rated}' : ' • ${context.l10n.casual}';
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: [
-        Icon(
-          seek.perf.icon,
-          color: DefaultTextStyle.of(context).style.color,
-        ),
-        const SizedBox(width: 4.0),
-        Text('${seek.timeIncrement.display}$mode'),
-      ],
     );
   }
 }
